+++ KinKal/StrawHit.hh
@@ -18,7 +18,7 @@ namespace KinKal {
       typedef std::shared_ptr<STRAWXING> STRAWXINGPTR;
       StrawHit(BField const& bfield, TLine const& straj, D2T const& d2t, STRAWXINGPTR const& sxing,LRAmbig ambig=LRAmbig::null) :
 	WHIT(sxing, bfield,straj,d2t,sxing->strawMat().strawRadius(),ambig) {}
-      virtual float tension() const override { return 0.0; } // check against straw diameter, length, any other measurement content FIXME!
+      virtual double tension() const override { return 0.0; } // check against straw diameter, length, any other measurement content FIXME!
       virtual void print(std::ostream& ost=std::cout,int detail=0) const override;
       virtual ~StrawHit(){}
     private:
